Refactoring the code (Changes)
