+++ lib/business/Expenses/all_expenses_screen/all_expenses_list.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:kuza/business/Expenses/models/database_provider.dart';
+import 'package:kuza/data/database_provider.dart';
 import 'package:provider/provider.dart';
 import '../expense_screen/expense_card.dart';
