+++ lib/business/Expenses/category_screen/category_card.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
-import 'package:kuza/business/Expenses/models/ex_category.dart';
+import 'package:kuza/models/ex_category.dart';
 import 'package:kuza/business/Expenses/screens/expense_screen.dart';
 
 class CategoryCard extends StatelessWidget {
