+++ lib/business/Expenses/category_screen/category_fetcher.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:kuza/business/Expenses/models/database_provider.dart';
+import 'package:kuza/data/database_provider.dart';
 import 'package:kuza/business/Expenses/screens/all_expenses.dart';
 import 'package:provider/provider.dart';
 import './total_chart.dart';
