+++ lib/business/Expenses/expense_form.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
 import 'package:kuza/business/Expenses/constants/icons.dart';
-import 'package:kuza/business/Expenses/models/database_provider.dart';
-import 'package:kuza/business/Expenses/models/expense.dart';
+import 'package:kuza/data/database_provider.dart';
+import 'package:kuza/models/expense.dart';
 import 'package:provider/provider.dart';
 
 class ExpenseForm extends StatefulWidget {
