+++ lib/business/Expenses/expense_screen/expense_chart.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
-import 'package:kuza/business/Expenses/models/database_provider.dart';
+import 'package:kuza/data/database_provider.dart';
 import 'package:provider/provider.dart';
 import 'package:fl_chart/fl_chart.dart';
