+++ lib/business/Expenses/widgets/all_expenses_screen/all_expenses_fetcher.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import '../../models/database_provider.dart';
+import '../../../../data/database_provider.dart';
 import './all_expenses_list.dart';
 import './expense_search.dart';
