+++ lib/business/Expenses/widgets/all_expenses_screen/all_expenses_list.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import '../../models/database_provider.dart';
+import '../../../../data/database_provider.dart';
 import '../expense_screen/expense_card.dart';
 
 class AllExpensesList extends StatelessWidget {
