+++ lib/business/Expenses/widgets/all_expenses_screen/expense_search.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import '../../models/database_provider.dart';
+import '../../../../data/database_provider.dart';
 
 class ExpenseSearch extends StatefulWidget {
   const ExpenseSearch({super.key});
