+++ lib/business/Expenses/widgets/category_screen/category_card.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
-import '../../models/ex_category.dart';
+import '../../../../models/ex_category.dart';
 import '../../screens/expense_screen.dart';
 
 class CategoryCard extends StatelessWidget {
