+++ lib/business/Expenses/widgets/category_screen/category_fetcher.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import '../../models/database_provider.dart';
+import '../../../../data/database_provider.dart';
 import './total_chart.dart';
 import './category_list.dart';
 import '../../screens/all_expenses.dart';
