+++ lib/business/Expenses/widgets/category_screen/category_list.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import '../../models/database_provider.dart';
+import '../../../../data/database_provider.dart';
 import './category_card.dart';
 
 class CategoryList extends StatelessWidget {
