+++ lib/business/Expenses/widgets/expense_screen/expense_card.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
-import '../../models/expense.dart';
+import '../../../../models/expense.dart';
 import '../../constants/icons.dart';
 import './confirm_box.dart';
