+++ lib/business/Expenses/widgets/expense_screen/expense_chart.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
 import 'package:provider/provider.dart';
 import 'package:fl_chart/fl_chart.dart';
-import '../../models/database_provider.dart';
+import '../../../../data/database_provider.dart';
 
 class ExpenseChart extends StatefulWidget {
   final String category;
