+++ lib/main.dart
@@ -1,6 +1,6 @@
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
-import 'package:kuza/business/Expenses/models/database_provider.dart';
+import 'package:kuza/data/database_provider.dart';
 import 'package:kuza/business/Expenses/screens/all_expenses.dart';
 import 'package:kuza/business/Point%20of%20sale%20(POS)/cart_model.dart';
 import 'package:kuza/pages/brand_intro.dart';
